API/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
geopandas
openpyxl
Fiona
Shapely
python-dotenv
mysql-connector-python